cnp - Add scaffolding for tests
