+++ test/flows/image_selenium_test.rb
@@ -1,0 +1,4 @@
+require '/Users/ciarra.peters/image-sharer-ciarrapeters/test/test_helper.rb'
+
+class ImageTestSelenium < ActionDispatch::IntegrationTest
+end
