+++ test/page_objects/images/index_page.rb
@@ -1,0 +1,5 @@
+module Images
+  class IndexPage < AePageObjects::Document
+    path :root
+  end
+end
