+++ test/page_objects/images/new_page.rb
@@ -1,0 +1,5 @@
+module Images
+  class NewPage < AePageObjects::Document
+    path :new_image
+  end
+end
