+++ test/page_objects/images/show_page.rb
@@ -1,0 +1,5 @@
+module Images
+  class ShowPage < AePageObjects::Document
+    path :image
+  end
+end
